program to remove a word from string
In this program we input a string (sentence) from the user and accept the word position of the word which is to be deleted. If the position is an invalid one then suitable message will be displayed else the modified sentence will be shown. The outputs are also given at the end of the code.